rename network_get network_put modules to net_get net_put (#40381)
