+++ test/integration/targets/ios_file/tests/cli/net_get.yaml
@@ -1,5 +1,5 @@
 ---
-- debug: msg="START ios cli/network_get.yaml on connection={{ ansible_connection }}"
+- debug: msg="START ios cli/net_get.yaml on connection={{ ansible_connection }}"
 
 # Add minimal testcase to check args are passed correctly to
 # implementation module and module run is successful.
@@ -13,7 +13,7 @@
     match: none
 
 - name: setup (copy file to be fetched from device)
-  network_put:
+  net_put:
     src: ios1.cfg
   register: result
 
@@ -22,7 +22,7 @@
       - result.changed == true
 
 - name: get the file from device with dest unspecified
-  network_get:
+  net_get:
     src: ios1.cfg
   register: result
 
@@ -31,7 +31,7 @@
       - result.changed == true
 
 - name: get the file from device with relative destination
-  network_get:
+  net_get:
     src: ios1.cfg
     dest: 'ios_{{ ansible_host }}.cfg'
   register: result
@@ -40,4 +40,4 @@
     that:
       - result.changed == true
 
-- debug: msg="END ios cli/network_get.yaml on connection={{ ansible_connection }}"
+- debug: msg="END ios cli/net_get.yaml on connection={{ ansible_connection }}"
